small refactor
Move check_dataarray_coords in xarray.core.coordinates module and rename
it to validate_dataarray_coords (name consistent with
Index.validate_dataarray_coord).

Move CoordinateValidationError from xarray.core.indexes to
xarray.core.coordinates module.
